Remove datakat header from roadobjectclient
This breaks the API since the public method `getDatakatalog` is removed
from RoadObjectClient. However, when this method is deleted there is no
furter use for datakatalog for RoadObjectClient and no other clients are
sending the datakat header. Therefore the sending for datakat header is
removed, but this breaks the semantics for the client since it will no
longer be able to detect datakatalog version changes.